requirements.txt
Pillow
RPi.GPIO
spidev
numpy

# sudo apt-get -y install libatlas-base-dev